+++ product_front/app/src/main/java/com/example/product_front/ScanScreen.kt
@@ -1,0 +1,37 @@
+package com.example.product_front
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+fun ScanScreen() {
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = MaterialTheme.colorScheme.background
+    ) {
+        Column(
+            modifier = Modifier.fillMaxSize(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            // NOTE: You will need to integrate a real camera view and permissions
+            Text(text = "Camera View Placeholder")
+            Spacer(modifier = Modifier.height(16.dp))
+            Button(onClick = { /* Implement the photo capture functionality */ }) {
+                Text(text = "Capture")
+            }
+        }
+    }
+}
